+++ src/Ui/Bundle/EventBundle/Resources/assets/js/components/form/_AddSubmitEventListenerOnElementChange.js
@@ -1,0 +1,15 @@
+function addSubmitEventListenerOnElementChange(target, formName, elementName) {
+    const form = target.querySelector(`form[name="${formName}"]`);
+
+    if (null !== form) {
+        form.querySelectorAll(`[name="${formName}[${elementName}]"]`)
+            .forEach(function (elementInput) {
+                elementInput.addEventListener('change', function () {
+                    form.submit();
+                });
+            })
+        ;
+    }
+}
+
+module.exports = addSubmitEventListenerOnElementChange;