+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- VIMEET-2102 - Evaluate visio meeting
+- VIMEET-2128 - Show participant name on visio
+
+### Fixed
+- MV-236 - Add unavailability with a different timezone
+- Add number validator for priority meeting request number
 
 ## [2.31.0] - 2020-06-01
 ### Added